@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.WebSocket;
+using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.IO;
 using System.Threading.Tasks;
@@ -8,13 +9,13 @@ namespace Marina.Utils
 {
     public static class Console
     {
-        private static readonly FileInfo LogFile = new FileInfo("Marina.log");
+        private static readonly FileInfo LogFile = new("Marina.log");
 
         static Console()
         {
-            Program.Initialize += delegate (object? sender, DiscordSocketClient client)
+            Program.Initialize += delegate (object? sender, ServiceProvider services)
             {
-                client.Log += async delegate (LogMessage log)
+                services.GetService<DiscordSocketClient>().Log += async delegate (LogMessage log)
                 {
                     await WriteLog($"[{DateTime.Now}]: {log.ToString()}\n");
                 };
